Updating copyfile api
This update includes a newer promise based system for our file system plugin.  This simplifies code and makes it easier for maintainability.  More promise based systems will come.

Added a checking procedure to see if the current payroll has been finalized before overwriting it with new data.  If it has not been finalized, it will go through the finalization process.